treemanifest: support readdelta in fastmanifest
If we were using fastmanifest in a treeonly repository, hg show would try to
access the manifest delta via the revlog which would crash. The fix is to make
fastmanifest first try accessing the data via the tree, before falling back to
the delta.

My attempts at automatically testing this failed, but I had a repro in a real
repository and confirmed this fixed the issue.