treemanifest: catch remote output during tree fetching
There was a flakey test because the remote stderr output was getting eaten part
of the time. This fixes it by capturing all the output and then sending it to
stderr manually.

Tested it by running all the tests in parallel multiple times and not seeing the
issue anymore.

Also we now send that output to the debug output. We're already throwing a more
useful error message, so let's hide the noise.

Differential Revision: https://phab.mercurial-scm.org/D1276