SingleNodeHash revset should not fail if the hash is unknown
Summary:
It's not useful to have revsets fail when the revision is unknown -
instead, make `SingleNodeHash` return 0 or 1 revisions, depending on whether
the node is known. We can add back the failure behaviour if we find a use for
it - e.g. by adding a filter revset that returns an error if it does not see a
given node passing through.

Reviewed By: jsgf

Differential Revision: D6239830

fbshipit-source-id: 797de26bedfa52f2edf0c5a3b5865866e7567d2a