basepack: add a lru cache for the pack files
Summary:
To speed up pack lookups (especially when there are lots of packs), we
should maintain an lru ordering of the packs and perform searches in that
order, since it's likely the next entry we search for will be in the same pack
file as the last entry we searched for. This commit achieves the same.

Test Plan:
 - Ran all the tests.

 - Created ~2k pack files in a large repo.

  - Time taken without the cache:
    - `hg update b` while at a: ~18 minutes.
    - `hg update a` while at b: ~23 seconds.

  - Time taken with the cache:
    - `hg update b` while at a: ~14 seconds.
    - `hg update a` while at b: ~9 seconds.

Reviewers: #fbhgext, durham

Reviewed By: #fbhgext, durham

Subscribers: durham

Differential Revision: https://phab.mercurial-scm.org/D1208