Make storage_types::Version's field public
Summary:
Implementations of various storage backends in Mononoke may need to access the raw value inside of a Version struct. Originally, the intention was to make Version a trait, with concrete implementations inside of each storage backend crate that could access private fields by virtue of being in the same crate. Now that we're using a single concrete Version implementation across all of Mononoke, this field needs to be made public to allow the storage backends to have the appropriate access.

I considered several alternatives here: for example, we could instead provide as `as_inner()` method to extract the value, or we could even make this method part of a separate trait that one needs to specifically import to get access to the Version's innards. But given that all of these approaches still amount to allowing public access to this field, I decided to start out with the simplest approach, and ask: //should we instead make this a method?//

Reviewed By: jsgf

Differential Revision: D6215854

fbshipit-source-id: b29e7f9c96b077a5805b7195f28ddb5d871dca35