Add Core to test harness in bookmarks-test
Summary:
The `Bookmarks` trait assumes that the underlying implementation is asynchronous (as evidenced by the fact that all methods return `Future`s). Most non-toy implementations of an asynchronous bookmark store will need to schedule arbitrary tasks on the Core executing the returned Futures. (In some cases, the constructor itself may need to schedule tasks on the Core.) As such, this test harness should expose a Core to each test in a first class way (rather than trying to shoehorn it in via the `state`, which doesn't actually quite work here).

Existing implementations can just ignore the Core.

Reviewed By: jsgf

Differential Revision: D6215855

fbshipit-source-id: ea3f28bdaec667111b384db29d0c7c49954398e5