Use NodeHashes instead of Strings for values in bookmarks-test
Summary: In practice, Mononoke requires that the Value associated type of the Bookmarks trait be NodeHash. As a result, in the case of DbBookmarks, rather than making the implementation fully generic over values, I've made it specialized for NodeHashes for simplicity's sake. This means in order to use these tests, they need to store NodeHashes instead of Strings.

Reviewed By: farnz

Differential Revision: D6217789

fbshipit-source-id: 3f0f04180eb0869eb9de3da7be634d95c2b24d84