treemanifest: improve server side incremental repack
Currently an incremental repack on the server will repack the entire pack files,
and the new parts of the revlogs. The pack files can be very large and can take
a long, long time to run. So let's use the normal incremental pack heuristics to
minimize how often we have to do full repacks.

Differential Revision: https://phab.mercurial-scm.org/D1350