treemanifest: support "{manifest % '{node}'}" template
When we changed the treemanifest {manifest} template output it broke the ability
to specifically ask for the node. This is important for tools migrating between
the old and new format. Let's add that back in.

Let's also make tweakdefaults change the '{manifest}' default template for all
repo's, not just tree repos.

Differential Revision: https://phab.mercurial-scm.org/D1418