add a flag to hg_import_helper to show a manifest node ID
Summary:
Add a new debugging flag to the hg_import_helper.py script to have it just show
the manifest node ID for a given revision and then exit.

This is information is useful when debugging import issues, and I'm not aware
of an easy way to get it from the normal hg command line.

Reviewed By: wez

Differential Revision: D6332803

fbshipit-source-id: 600feac105f277d398325d2e3c50eae3d5200940