update to look for new error message thrown by treemanifest
Summary:
https://phab.mercurial-scm.org/D1268 changed the error message that the
treemanifest code throws when it cannot fetch tree data from the remote server.

This unfortunately is thrown as a generic `error.Abort()` so eden has to match
on the error message to tell if this was an error fetching the tree vs
something else.  Changing the error message broke eden's detection of this.

This updates the string that eden searches for.  In the future it would
probably be nicer to make the treemanifest code throw a different exception
type so we can detect this from the exception type rather than the error
message.

Reviewed By: wez

Differential Revision: D6332802

fbshipit-source-id: 629097914e6f783729adcd25f2fd5b6d96886ab4