improve error message when trying to re-clone a repo
Summary:
The messaging implies that something is already mounted, especially
if one is used to the error messages that are emitted by `mount`.  In this
situation we're complaining that we already have a repo configured at that
path, so let's help point folks in the right direction in the error message.

Reviewed By: chadaustin

Differential Revision: D6344017

fbshipit-source-id: d48176f5806b3233e191567a840c88af101002cf