tests: fix too many files error in datapack store
If the os limited a given process to a <200 files open, this test would fail.
Let's change the cache size to be smaller to avoid this.

Also, it turns out the cache size and number of packs created doesn't actually
seem to affect this test. I changed the numbers in a few ways and the test never
failed.

Differential Revision: https://phab.mercurial-scm.org/D1435