update EdenServer to implement TakeoverHandler
Summary:
Update EdenServer to implement the TakeoverHandler API, and to exit after
sending the mount point takeover data.  The actual shutdown logic itself is not
implemented yet--this just sends dummy data for now.  However, this does serve
as a proof of concept that the TakeoverServer and TakeoverClient code functions
as desired.

Reviewed By: bolinfest

Differential Revision: D6018180

fbshipit-source-id: c19581928926a46b767f1ee5c1761381e5055fa9