Directory was showing up as UNTRACKED_ADDED instead of its contents.
Summary:
This bug is part of a bigger issue in our Mercurial integration where
`UNTRACKED_ADDED` conflicts are being silently swallowed in our Hg extension
whereas stock Mercurial presents these as conflicts and forces the user to deal
with them. The Mercurial issues will be addressed in a follow-up change.

Reviewed By: simpkins

Differential Revision: D6365580

fbshipit-source-id: 831e27ce1da90ea605033b2b9988fe400ba404aa