Remove excludes from a build rule that no longer exist.
Summary:
This should have been removed as part of D6179950.

Ideally, Buck would error when this happens, but apparently `glob()` does not
complain when patterns do not match any files, even when the pattern does not
contain any wildcards. There appears to be some code at Facebook that is
exploiting this behavior.

Reviewed By: simpkins

Differential Revision: D6421529

fbshipit-source-id: c6f982624e0e12a911bc12ab1e8239ba4358ea56