treedirstate: implement casefolding maps for case insensitive filesystems
The dirfoldmap and filefoldmap on the dirstatemap object map from normalized
filename to the form it appears in the dirstate (denormalized form).

The dicts these functions return are also modified by the dirstate with
filenames discovered by examining the disk, so use an overlay dictionary backed
by the real dirstate to allow this to happen.

Differential Revision: https://phab.mercurial-scm.org/D1409