treedirstate: better iteration using visitor pattern
In order to allow the python interface code to perform actions on each node in
the tree without creating python-specific interfaces in the generic Rust code,
add a method of iterating over the tree, executing a closure at each file.

Use this to implement the methods that give iterators over the filenames in the
tree.  This performs better than the get_first/get_next-style iterators.

Differential Revision: https://phab.mercurial-scm.org/D1403