treedirstate: implement efficient case collision detection
Add a mechanism to the dirstate trees to allow lookups based on filtered views
of the keys.  For a given filtering function, this returns one (if any) of the
keys for which filter(key) matches the input.  The filtered values in each
directory node are cached to improve subsequent lookups.

Differential Revision: https://phab.mercurial-scm.org/D1406