remotefilelog: make sure metacache is always initalized
This adds initialization of metacache in remaining place where it can be uninitalized when using threads
The previous state did fail on some hg update operations on Windows. It didn't
affect posix as that doesn't actually use threads so we always have metacache
initalized from the constructor

Test Plan:
Run hg update that failed wihtout this change on windows and see it
suceed.
run tests on CentOS

Differential Revision: https://phab.mercurial-scm.org/D1539