fix state handling in EdenMount::destroy()
Summary:
This fixes a crash in EdenMount::destroy() if EdenMount::create() failed to
load the root inode.  Previously the code called shutdownImpl() in this case
which tried to unload all inodes and crashed since the root inode was null.

This also fixes EdenMount::destroy() to properly handle the FUSE_ERROR and
FUSE_DONE cases.

Reviewed By: wez

Differential Revision: D6434355

fbshipit-source-id: 39c5f4472d6ebbcf881b4c9c8c8fd67686032ec1