treemanifest: revert e64038f8 which used the connectionpool for pull
The patch that tried to reuse a connection during pull was actually fairly
broken. For one, if multiple peers were created during a pull it would reuse the
same peer, even if the other peer was still in use. Another issue is that hg
pull closes its peer when it's finished, which would return a bad connection
into the pool.

Let's just revert this entirely for now, since there's not an easy way to do
this without refactoring in core.